refactor(hr): convert component to typescript
