+++ src/components/hr/hr.style.ts
@@ -1,8 +1,8 @@
 import styled from "styled-components";
-import { margin } from "styled-system";
+import { margin, MarginProps } from "styled-system";
 import baseTheme from "../../style/themes/base";
 
-const StyledHr = styled.hr`
+const StyledHr = styled.hr<MarginProps>`
   ${margin}
   width: inherit;
   border: 0;
